# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lotus Pharmacies CRM & Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  agent: "agent",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  agent: "agent",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  agent: "agent",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
  password?: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  branch?: string;
  tags?: string[];
  notes?: string;
  prescriptionNotes?: string;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  phone: string;
  branch?: string;
  tags?: string[];
  notes?: string;
  prescriptionNotes?: string;
}

export interface UpdateCustomerBody {
  name?: string;
  phone?: string;
  branch?: string;
  tags?: string[];
  notes?: string;
  prescriptionNotes?: string;
}

export type ConversationStatus =
  (typeof ConversationStatus)[keyof typeof ConversationStatus];

export const ConversationStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
} as const;

export interface Conversation {
  id: number;
  customerId: number;
  customer?: Customer;
  assignedAgentId?: number | null;
  assignedAgent?: User | null;
  status: ConversationStatus;
  tags: string[];
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
  resolvedAt?: string | null;
}

export interface CreateConversationBody {
  customerId: number;
  tags?: string[];
  assignedAgentId?: number | null;
}

export type UpdateConversationBodyStatus =
  (typeof UpdateConversationBodyStatus)[keyof typeof UpdateConversationBodyStatus];

export const UpdateConversationBodyStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
} as const;

export interface UpdateConversationBody {
  tags?: string[];
  status?: UpdateConversationBodyStatus;
  assignedAgentId?: number | null;
}

export interface AssignConversationBody {
  agentId: number | null;
}

export type MessageSenderType =
  (typeof MessageSenderType)[keyof typeof MessageSenderType];

export const MessageSenderType = {
  agent: "agent",
  customer: "customer",
  system: "system",
} as const;

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus];

export const MessageStatus = {
  sent: "sent",
  delivered: "delivered",
  read: "read",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  senderId?: number | null;
  senderType: MessageSenderType;
  body: string;
  attachments?: string[];
  status: MessageStatus;
  isNote: boolean;
  createdAt: string;
}

export type SendMessageBodySenderType =
  (typeof SendMessageBodySenderType)[keyof typeof SendMessageBodySenderType];

export const SendMessageBodySenderType = {
  agent: "agent",
  customer: "customer",
  system: "system",
} as const;

export interface SendMessageBody {
  body: string;
  senderType: SendMessageBodySenderType;
  isNote?: boolean;
  attachments?: string[];
}

export interface Tag {
  id: number;
  name: string;
  color: string;
}

export interface CreateTagBody {
  name: string;
  color: string;
}

export interface QuickReply {
  id: number;
  title: string;
  body: string;
  createdAt: string;
}

export interface CreateQuickReplyBody {
  title: string;
  body: string;
}

export interface DashboardStats {
  totalConversations: number;
  activeChats: number;
  resolvedChats: number;
  avgResponseTimeMinutes: number;
  openConversations: number;
  newConversationsToday: number;
  totalCustomers: number;
  totalAgents: number;
}

export interface ChatsOverTimePoint {
  date: string;
  count: number;
}

export interface AgentPerformance {
  agentId: number;
  agentName: string;
  totalHandled: number;
  resolved: number;
  avgResponseTimeMinutes: number;
  activeChats: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  conversation_created: "conversation_created",
  conversation_resolved: "conversation_resolved",
  message_sent: "message_sent",
  agent_assigned: "agent_assigned",
  customer_created: "customer_created",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  agentName?: string | null;
  customerName?: string | null;
  createdAt: string;
}

export type ListCustomersParams = {
  search?: string;
  tag?: string;
  branch?: string;
};

export type ListConversationsParams = {
  status?: ListConversationsStatus;
  agentId?: number;
  tag?: string;
  search?: string;
};

export type ListConversationsStatus =
  (typeof ListConversationsStatus)[keyof typeof ListConversationsStatus];

export const ListConversationsStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
} as const;